requirements.txt
Django>=2.1
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
django-cors-headers
